+++ R/conversion.R
@@ -14,6 +14,7 @@
 #'     default it operates with bare expressions, but this
 #'     behavior can be disabled by a specifying \code{mode = "standard"} or setting
 #'     \code{units_options(set_units_mode = "standard")}.
+#'     If \code{value} is missing or set to \code{1}, the object becomes unitless.
 #' }
 #'
 #' @param x numeric vector, or object of class \code{units}.
@@ -193,12 +194,14 @@ as_difftime <- function(x) {
 
 
 #' @export
-`[.units` <- function(x, i, j, ..., drop = TRUE)
-  .as.units(NextMethod(), units(x))
+`[.units` <- function(x, i, j, ..., drop = TRUE) {
+  restore_units(NextMethod(), x)
+}
 
 #' @export
-`[[.units` <- function(x, i, j, ...)
-  .as.units(NextMethod(), units(x))
+`[[.units` <- function(x, i, j, ...) {
+  restore_units(NextMethod(), x)
+}
 
 #' @export
 as.POSIXct.units = function (x, tz = "UTC", ...) {