+++ R/tidyverse.R
@@ -6,12 +6,11 @@ type_sum.units <- function(x, ...) {
 }
 
 type_sum.mixed_units <- function(x, ...) {
-  "mixed_units" 
+  "mixed_units"
 }
 
 pillar_shaft.units <- function(x, ...) {
-  out <- format(unclass(x), ...)
-  pillar::new_pillar_shaft_simple(out, align = "right", min_width = 8)
+  pillar::pillar_shaft(unclass(x), ...)
 }
 
 pillar_shaft.mixed_units <- function(x, ...) {
@@ -31,7 +30,12 @@ vec_proxy.units = function(x, ...) {
   x
 }
 vec_restore.units = function(x, to, ...) {
-  set_units(x, units(to), mode = "standard")
+  restore_units(x, to)
+}
+restore_units <- function(x, to) {
+  out <- .as.units(x, units(to))
+  attr(out, "pillar") <- attr(to, "pillar")
+  out
 }