+++ tests/testthat/test_conversion.R
@@ -1,5 +1,3 @@
-context("Unit conversion")
-
 test_that("we can convert numbers to unit-less units", {
   x <- as_units(1:4)
   expect_equal(length(x), 4)
@@ -150,11 +148,11 @@ test_that("conversion to dimensionless with prefix works (g/kg) if simplify=TRUE
 	units(a) = as_units("mg/kg")
 	expect_equal(as.numeric(a), a_orig/1e6)
 	units(a) = as_units("kg/mg")
-	expect_equal(a, a_orig)
+	expect_equal(a, set_units(a_orig, 1))
 	units(a) = as_units("g/g")
-	expect_equal(a, a_orig)
+	expect_equal(a, set_units(a_orig, 1))
 	units(a) = as_units("kg/g")
-	expect_equal(a, a_orig * 1000)
+	expect_equal(a, set_units(a_orig * 1000, 1))
 	units_options(simplify = NA)
 })
 
@@ -201,7 +199,7 @@ test_that("errors are correctly coerced to a data frame", {
 test_that("units are correctly coerced to a list", {
   x <- 1:10 * as_units("m")
   y <- as.list(x)
-  expect_is(y, "list")
+  expect_type(y, "list")
   expect_true(all(sapply(seq_along(y), function(i) all.equal(y[[i]], x[i]))))
 })
 
@@ -212,10 +210,9 @@ test_that("NA as units generate warnings", {
 
 test_that("ud_are_convertible return the expected value", {
   x <- 1:10 * as_units("m")
-  expect_is(ud_are_convertible("m", "km"), "logical")
+  expect_type(ud_are_convertible("m", "km"), "logical")
   expect_true(ud_are_convertible("m", "km"))
   expect_true(ud_are_convertible(units(x), "km"))
   expect_false(ud_are_convertible("s", "kg"))
 
 })
-